initial cleanup moved AnnualReport to /views includes auth token in main.ts
