+++ coops-ui/src/assets/styles/base.styl
@@ -1,4 +1,4 @@
-@import "theme.styl";
+@import "./theme.styl"
 
 html, body
   min-height 100vh
