+++ coops-ui/src/assets/styles/layout.styl
@@ -68,7 +68,8 @@ $app-header-height = 70px
     flex-flow row nowrap
 
     article
-      margin-right 2rem
+      + aside
+        margin-left 2rem
 
     aside
       margin-top 0
