+++ coops-ui/src/assets/styles/overrides.styl
@@ -29,3 +29,11 @@
 .v-card__text
   padding 1.5rem
   font-weight 300
+
+// Expansion Panel
+.v-expansion-panel__header
+  padding 1.25rem 1.25rem
+
+.v-expansion-panel__body
+  padding-right 1.25rem
+  padding-left 1.25rem
