+++ coops-ui/src/views/AnnualReport.vue
